requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#	production that isn't in development.
-r base.txt

gunicorn==18.0
django-storages==1.1.8
Collectfast==0.1.13
gevent==1.0
boto==2.25.0
python-memcached==1.53
#psycopg2
django-pylibmc-sasl==0.2.4